requirements.txt
python-docx
transformers
python-dotenv
